# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cross-device alarm sync API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Pair {
  id: number;
  code: string;
  createdAt: string;
}

export type DeviceRole = (typeof DeviceRole)[keyof typeof DeviceRole];

export const DeviceRole = {
  primary: "primary",
  secondary: "secondary",
} as const;

export interface Device {
  id: number;
  pairId: number;
  role: DeviceRole;
  createdAt: string;
}

export type DeviceInputRole =
  (typeof DeviceInputRole)[keyof typeof DeviceInputRole];

export const DeviceInputRole = {
  primary: "primary",
  secondary: "secondary",
} as const;

export interface DeviceInput {
  pairCode: string;
  role: DeviceInputRole;
}

export interface Alarm {
  id: number;
  pairId: number;
  label: string;
  hour: number;
  minute: number;
  enabled: boolean;
  createdAt: string;
}

export interface AlarmInput {
  pairCode: string;
  label: string;
  /**
   * @minimum 0
   * @maximum 23
   */
  hour: number;
  /**
   * @minimum 0
   * @maximum 59
   */
  minute: number;
}

export interface AlarmUpdate {
  label?: string;
  /**
   * @minimum 0
   * @maximum 23
   */
  hour?: number;
  /**
   * @minimum 0
   * @maximum 59
   */
  minute?: number;
  enabled?: boolean;
}

export type ListAlarmsParams = {
  pairCode: string;
};