# lib/api-zod/src/generated/types/alarm.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cross-device alarm sync API
 * OpenAPI spec version: 0.1.0
 */

export interface Alarm {
  id: number;
  pairId: number;
  label: string;
  hour: number;
  minute: number;
  enabled: boolean;
  createdAt: string;
}