# lib/api-zod/src/generated/types/alarmUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cross-device alarm sync API
 * OpenAPI spec version: 0.1.0
 */

export interface AlarmUpdate {
  label?: string;
  /**
   * @minimum 0
   * @maximum 23
   */
  hour?: number;
  /**
   * @minimum 0
   * @maximum 59
   */
  minute?: number;
  enabled?: boolean;
}