# lib/api-zod/src/generated/types/device.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cross-device alarm sync API
 * OpenAPI spec version: 0.1.0
 */
import type { DeviceRole } from "./deviceRole";

export interface Device {
  id: number;
  pairId: number;
  role: DeviceRole;
  createdAt: string;
}