# lib/api-zod/src/generated/types/deviceInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cross-device alarm sync API
 * OpenAPI spec version: 0.1.0
 */
import type { DeviceInputRole } from "./deviceInputRole";

export interface DeviceInput {
  pairCode: string;
  role: DeviceInputRole;
}