# lib/api-zod/src/generated/types/pair.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cross-device alarm sync API
 * OpenAPI spec version: 0.1.0
 */

export interface Pair {
  id: number;
  code: string;
  createdAt: string;
}